src/gl/object.cpp: actually translate and rotate our objects using shaders properly!
 - we need to call BindUniform every time we need to update the uniform on the shader side
 - we also convert all usages of double to float as I am sick of that shit

+++ include/gl/model.h
@@ -18,7 +18,7 @@ class GLModel : public GLObject {
         
     public:
         
-        GLModel(char *fileName, double x, double y, double z);
+        GLModel(char *fileName, float x, float y, float z);
         ~GLModel();
 };
