+++ src/gl/rectprism.cpp
@@ -7,7 +7,7 @@ using namespace LSE;
 Initialize our prism. Store size attributes
 and allocate all necessary memory.
 */
-GLRectPrism::GLRectPrism(double x, double y, double z, double w, double h, double d) : GLPrimitive(x, y, z) {
+GLRectPrism::GLRectPrism(float x, float y, float z, float w, float h, float d) : GLPrimitive(x, y, z) {
     
     width = w;
     height = h;
@@ -24,134 +24,134 @@ void GLRectPrism::CalcVertices() {
     // front face
     
     // v0
-    vertices[0] = -width/2.0;
-    vertices[1] = -height/2.0;
-    vertices[2] = -depth/2.0;
+    vertices[0] = -width/2.0f;
+    vertices[1] = -height/2.0f;
+    vertices[2] = -depth/2.0f;
     
     // v1
-    vertices[3] = width/2.0;
-    vertices[4] = -height/2.0;
-    vertices[5] = -depth/2.0;
+    vertices[3] = width/2.0f;
+    vertices[4] = -height/2.0f;
+    vertices[5] = -depth/2.0f;
     
     // v2
-    vertices[6] = width/2.0;
-    vertices[7] = height/2.0;
-    vertices[8] = -depth/2.0;
+    vertices[6] = width/2.0f;
+    vertices[7] = height/2.0f;
+    vertices[8] = -depth/2.0f;
     
     // v3
-    vertices[9] = -width/2.0;
-    vertices[10] = height/2.0;
-    vertices[11] = -depth/2.0;
+    vertices[9] = -width/2.0f;
+    vertices[10] = height/2.0f;
+    vertices[11] = -depth/2.0f;
     
     // right face
     
     // v4
-    vertices[12] = width/2.0;
-    vertices[13] = -height/2.0;
-    vertices[14] = -depth/2.0;
+    vertices[12] = width/2.0f;
+    vertices[13] = -height/2.0f;
+    vertices[14] = -depth/2.0f;
     
     // v5
-    vertices[15] = width/2.0;
-    vertices[16] = -height/2.0;
-    vertices[17] = depth/2.0;
+    vertices[15] = width/2.0f;
+    vertices[16] = -height/2.0f;
+    vertices[17] = depth/2.0f;
     
     // v6
-    vertices[18] = width/2.0;
-    vertices[19] = height/2.0;
-    vertices[20] = depth/2.0;
+    vertices[18] = width/2.0f;
+    vertices[19] = height/2.0f;
+    vertices[20] = depth/2.0f;
     
     // v7
-    vertices[21] = width/2.0;
-    vertices[22] = height/2.0;
-    vertices[23] = -depth/2.0;
+    vertices[21] = width/2.0f;
+    vertices[22] = height/2.0f;
+    vertices[23] = -depth/2.0f;
     
     // top face
     
     // v8
-    vertices[24] = -width/2.0;
-    vertices[25] = height/2.0;
-    vertices[26] = -depth/2.0;
+    vertices[24] = -width/2.0f;
+    vertices[25] = height/2.0f;
+    vertices[26] = -depth/2.0f;
     
     // v9
-    vertices[27] = width/2.0;
-    vertices[28] = height/2.0;
-    vertices[29] = -depth/2.0;
+    vertices[27] = width/2.0f;
+    vertices[28] = height/2.0f;
+    vertices[29] = -depth/2.0f;
     
     // v10
-    vertices[30] = width/2.0;
-    vertices[31] = height/2.0;
-    vertices[32] = depth/2.0;
+    vertices[30] = width/2.0f;
+    vertices[31] = height/2.0f;
+    vertices[32] = depth/2.0f;
     
     // v11
-    vertices[33] = -width/2.0;
-    vertices[34] = height/2.0;
-    vertices[35] = depth/2.0;
+    vertices[33] = -width/2.0f;
+    vertices[34] = height/2.0f;
+    vertices[35] = depth/2.0f;
     
     // left face
     
     // v12
-    vertices[36] = -width/2.0;
-    vertices[37] = -height/2.0;
-    vertices[38] = depth/2.0;
+    vertices[36] = -width/2.0f;
+    vertices[37] = -height/2.0f;
+    vertices[38] = depth/2.0f;
     
     // v13
-    vertices[39] = -width/2.0;
-    vertices[40] = -height/2.0;
-    vertices[41] = -depth/2.0;
+    vertices[39] = -width/2.0f;
+    vertices[40] = -height/2.0f;
+    vertices[41] = -depth/2.0f;
     
     // v14
-    vertices[42] = -width/2.0;
-    vertices[43] = height/2.0;
-    vertices[44] = -depth/2.0;
+    vertices[42] = -width/2.0f;
+    vertices[43] = height/2.0f;
+    vertices[44] = -depth/2.0f;
     
     // v15
-    vertices[45] = -width/2.0;
-    vertices[46] = height/2.0;
-    vertices[47] = depth/2.0;
+    vertices[45] = -width/2.0f;
+    vertices[46] = height/2.0f;
+    vertices[47] = depth/2.0f;
     
     // bottom face
     
     // v16
-    vertices[48] = -width/2.0;
-    vertices[49] = -height/2.0;
-    vertices[50] = depth/2.0;
+    vertices[48] = -width/2.0f;
+    vertices[49] = -height/2.0f;
+    vertices[50] = depth/2.0f;
     
     // v17
-    vertices[51] = width/2.0;
-    vertices[52] = -height/2.0;
-    vertices[53] = depth/2.0;
+    vertices[51] = width/2.0f;
+    vertices[52] = -height/2.0f;
+    vertices[53] = depth/2.0f;
     
     // v18
-    vertices[54] = width/2.0;
-    vertices[55] = -height/2.0;
-    vertices[56] = -depth/2.0;
+    vertices[54] = width/2.0f;
+    vertices[55] = -height/2.0f;
+    vertices[56] = -depth/2.0f;
     
     // v19
-    vertices[57] = -width/2.0;
-    vertices[58] = -height/2.0;
-    vertices[59] = -depth/2.0;
+    vertices[57] = -width/2.0f;
+    vertices[58] = -height/2.0f;
+    vertices[59] = -depth/2.0f;
     
     // back face
     
     // v20
-    vertices[60] = width/2.0;
-    vertices[61] = -height/2.0;
-    vertices[62] = depth/2.0;
+    vertices[60] = width/2.0f;
+    vertices[61] = -height/2.0f;
+    vertices[62] = depth/2.0f;
     
     // v21
-    vertices[63] = -width/2.0;
-    vertices[64] = -height/2.0;
-    vertices[65] = depth/2.0;
+    vertices[63] = -width/2.0f;
+    vertices[64] = -height/2.0f;
+    vertices[65] = depth/2.0f;
     
     // v22
-    vertices[66] = -width/2.0;
-    vertices[67] = height/2.0;
-    vertices[68] = depth/2.0;
+    vertices[66] = -width/2.0f;
+    vertices[67] = height/2.0f;
+    vertices[68] = depth/2.0f;
     
     // v23
-    vertices[69] = width/2.0;
-    vertices[70] = height/2.0;
-    vertices[71] = depth/2.0;
+    vertices[69] = width/2.0f;
+    vertices[70] = height/2.0f;
+    vertices[71] = depth/2.0f;
 }
 
 /*
