+++ src/util/terrain/3D.cpp
@@ -50,7 +50,7 @@ Terrain3D InterpNoise3D(float startX, float startY, float startZ, int width, int
                             
                             for(int d = 0; d < modDepth; d += sampleRate) {
                                 
-                                points[w][h][d] = fractal.get(startX + double(w), startY + float(h), startZ + double(d));
+                                points[w][h][d] = fractal.get(startX + float(w), startY + float(h), startZ + float(d));
                                 
                                 if(d == 0)
                                     --d;
@@ -229,28 +229,28 @@ Terrain3D GenTerrain3D(int seed, float startX, float startY, float startZ, int w
     f0.setNumOctaves(3);
 	f0.setFrequency(0.157);
 	f0.setGain(10.362);
-	f0.setLacunarity(double(sampleRates[0]));
+	f0.setLacunarity(float(sampleRates[0]));
 	
 	anl::CImplicitFractal f1(anl::FBM, anl::SIMPLEX, anl::CUBIC);
     f1.setSeed(seed);
     f1.setNumOctaves(3);
 	f1.setFrequency(0.872);
 	f1.setGain(2479.789);
-	f1.setLacunarity(double(sampleRates[1]));
+	f1.setLacunarity(float(sampleRates[1]));
 	
 	anl::CImplicitFractal f2(anl::FBM, anl::SIMPLEX, anl::CUBIC);
     f2.setSeed(seed);
     f2.setNumOctaves(3);
 	f2.setFrequency(12.444);
 	f2.setGain(37.556);
-	f2.setLacunarity(double(sampleRates[2]));
+	f2.setLacunarity(float(sampleRates[2]));
 	
 	anl::CImplicitFractal f3(anl::FBM, anl::SIMPLEX, anl::CUBIC);
     f3.setSeed(seed);
     f3.setNumOctaves(3);
 	f3.setFrequency(17.533);
 	f3.setGain(5357.891);
-	f3.setLacunarity(double(sampleRates[3]));
+	f3.setLacunarity(float(sampleRates[3]));
     
     Terrain3D p0 = InterpNoise3D(startX, startY, startZ, width, height, depth, sampleRates[0], low, high, f0);
     Terrain3D p1 = InterpNoise3D(startX, startY, startZ, width, height, depth, sampleRates[1], low, high, f1);
